Sources of a recorder
It's a pretty stupid thing, it just records microphone and mixers (speaker output) into two independent files.
It also captures screenshots every second, so that you can then merge it into a complex file using some other software.
Sound record is started at the same time with the same bitrate and samplerate, so they should be in a pretty good sync, although no other synchronization techniques are used, so it's not really promised.